refactor: StoryDetailView MVVM 분리
